refactor(client): flatten render data classes
