+++ client/src/main/java/net/lapidist/colony/client/render/MapRenderData.java
@@ -1,8 +1,6 @@
 package net.lapidist.colony.client.render;
 
 import com.badlogic.gdx.utils.Array;
-import net.lapidist.colony.client.render.data.RenderBuilding;
-import net.lapidist.colony.client.render.data.RenderTile;
 
 /**
  * Read-only data describing map entities for rendering.
