+++ client/src/main/java/net/lapidist/colony/client/render/RenderBuilding.java
@@ -1,4 +1,4 @@
-package net.lapidist.colony.client.render.data;
+package net.lapidist.colony.client.render;
 
 /** Render data object for a building. */
 public final class RenderBuilding {
