--- client/src/main/java/net/lapidist/colony/client/render/data/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Plain data objects consumed by client renderers.
- */
-package net.lapidist.colony.client.render.data;
