+++ client/src/main/java/net/lapidist/colony/client/renderers/MapTileCache.java
@@ -15,7 +15,7 @@ import net.lapidist.colony.client.util.CameraUtils;
 import net.lapidist.colony.client.core.io.ResourceLoader;
 import net.lapidist.colony.client.TileRotationUtil;
 import net.lapidist.colony.client.render.MapRenderData;
-import net.lapidist.colony.client.render.data.RenderTile;
+import net.lapidist.colony.client.render.RenderTile;
 import net.lapidist.colony.client.systems.CameraProvider;
 import net.lapidist.colony.components.GameConstants;
