+++ tests/src/test/java/net/lapidist/colony/client/renderers/MapTileCacheTest.java
@@ -12,8 +12,8 @@ import net.lapidist.colony.client.core.io.ResourceLoader;
 import net.lapidist.colony.client.render.MapRenderData;
 import net.lapidist.colony.client.render.MapRenderDataBuilder;
 import net.lapidist.colony.client.render.SimpleMapRenderData;
-import net.lapidist.colony.client.render.data.RenderTile;
-import net.lapidist.colony.client.render.data.RenderBuilding;
+import net.lapidist.colony.client.render.RenderTile;
+import net.lapidist.colony.client.render.RenderBuilding;
 import com.badlogic.gdx.utils.Array;
 import net.lapidist.colony.client.systems.CameraProvider;
 import com.badlogic.gdx.graphics.OrthographicCamera;
