+++ tests/src/test/java/net/lapidist/colony/tests/client/render/data/RenderBuildingTest.java
@@ -1,6 +1,6 @@
 package net.lapidist.colony.tests.client.render.data;
 
-import net.lapidist.colony.client.render.data.RenderBuilding;
+import net.lapidist.colony.client.render.RenderBuilding;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
