+++ tests/src/test/java/net/lapidist/colony/tests/client/renderers/ResourceRendererTest.java
@@ -9,10 +9,10 @@ import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.utils.viewport.ExtendViewport;
 import net.lapidist.colony.client.renderers.ResourceRenderer;
 import net.lapidist.colony.client.systems.CameraProvider;
-import net.lapidist.colony.client.render.data.RenderTile;
+import net.lapidist.colony.client.render.RenderTile;
 import net.lapidist.colony.client.render.MapRenderData;
 import net.lapidist.colony.client.render.SimpleMapRenderData;
-import net.lapidist.colony.client.render.data.RenderBuilding;
+import net.lapidist.colony.client.render.RenderBuilding;
 import net.lapidist.colony.client.systems.MapRenderDataSystem;
 import net.lapidist.colony.components.state.MapState;
 import net.lapidist.colony.tests.GdxTestRunner;
