+++ tests/src/test/java/net/lapidist/colony/tests/client/renderers/TileRendererTest.java
@@ -11,10 +11,10 @@ import net.lapidist.colony.client.renderers.TileRenderer;
 import net.lapidist.colony.base.BaseDefinitionsMod;
 import net.lapidist.colony.client.core.io.ResourceLoader;
 import net.lapidist.colony.client.systems.CameraProvider;
-import net.lapidist.colony.client.render.data.RenderTile;
+import net.lapidist.colony.client.render.RenderTile;
 import net.lapidist.colony.client.render.MapRenderData;
 import net.lapidist.colony.client.render.SimpleMapRenderData;
-import net.lapidist.colony.client.render.data.RenderBuilding;
+import net.lapidist.colony.client.render.RenderBuilding;
 import net.lapidist.colony.components.state.MapState;
 import net.lapidist.colony.tests.GdxTestRunner;
 import org.junit.Test;
